requirements.txt
mcp
pydantic
python-dotenv
openai
ipaddr